Fix conflict between merge and current branch
